Fixed Task3NonMonteCarlo
Fixed Task3NonMonteCarlo. The original solution was incorrect as it mistakenly considered players' scores to be independant. Sadly this made this solution even slower and I had to remove the average score tracking as in sufficiently long games (e.g. 1000 die rolls, sequence size of 3, die size of 6) it returned NaN instead of score.